+++ DigThemGraves/Assets/Prefabs/Inventory&Shop.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 590442cf73b2c1547be0ec48c4d65c16
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
