+++ DigThemGraves/Assets/ScriptableObjects/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 598968f1f3ebd784b9afbbb41171a364
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
