+++ DigThemGraves/Assets/ScriptableObjects/Items/Stone Shovel.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 112c65b020ab3504f8f13f990976cae5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
