--- DigThemGraves/Assets/Scripts/Money/Iron Shovel.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 887fa6adcee01534bb149b2ef4abda3f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
