--- DigThemGraves/Assets/Scripts/Money/ItemPanel.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: df3030facb8302f43bced60c6064f4eb
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
