--- DigThemGraves/Assets/Scripts/Money/Wood Shovel.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e796aa84abf263443a5b05f5b097c4b5
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
