+++ DigThemGraves/Assets/Scripts/MoneyInventoryShop.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9188c3dfb3d88d346831b7a48b990e3d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
