+++ DigThemGraves/Assets/Scripts/MoneyInventoryShop/ItemSlot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c4d074e008116894f87a86046c30e623
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
