+++ DigThemGraves/Assets/Scripts/MoneyInventoryShop/ItemTemplate.cs
