+++ DigThemGraves/Assets/Scripts/MoneyInventoryShop/ItemTemplate.cs.meta
