+++ DigThemGraves/Assets/Sprites/Temporary/MoneySprite.png
